+++ templates/with-vercel-postgres/next.config.ts
@@ -1,15 +1,12 @@
 import { withPayload } from '@payloadcms/next/withPayload'
-
+import type { NextConfig } from 'next'
 import path from 'path'
-
 import { fileURLToPath } from 'url'
 
 const __filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(__filename)
 
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  // Your Next.js config here
+const nextConfig: NextConfig = {
   webpack: (webpackConfig) => {
     webpackConfig.resolve.extensionAlias = {
       '.cjs': ['.cts', '.cjs'],
