+++ templates/with-vercel-website/next.config.ts
@@ -1,31 +1,31 @@
 import { withPayload } from '@payloadcms/next/withPayload'
+import type { NextConfig } from 'next'
 import path from 'path'
-
 import { fileURLToPath } from 'url'
 
 const __filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(__filename)
-import redirects from './redirects.js'
 
-const NEXT_PUBLIC_SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL || 'http://localhost:3000'
+import { redirects } from './redirects.js'
 
-/** @type {import('next').NextConfig} */
-const nextConfig = {
+const NEXT_PUBLIC_SERVER_URL = process.env.VERCEL_PROJECT_PRODUCTION_URL
+  ? `https://${process.env.VERCEL_PROJECT_PRODUCTION_URL}`
+  : process.env.__NEXT_PRIVATE_ORIGIN || 'http://localhost:3000'
+
+const nextConfig: NextConfig = {
   images: {
-    qualities: [90, 100],
+    qualities: [100],
     remotePatterns: [
       ...[NEXT_PUBLIC_SERVER_URL /* 'https://example.com' */].map((item) => {
         const url = new URL(item)
 
         return {
           hostname: url.hostname,
-          protocol: url.protocol.replace(':', ''),
+          protocol: url.protocol.replace(':', '') as 'http' | 'https',
         }
       }),
     ],
   },
-  reactStrictMode: true,
-  redirects,
   webpack: (webpackConfig) => {
     webpackConfig.resolve.extensionAlias = {
       '.cjs': ['.cts', '.cjs'],
@@ -35,9 +35,11 @@ const nextConfig = {
 
     return webpackConfig
   },
+  reactStrictMode: true,
+  redirects,
   turbopack: {
     root: path.resolve(dirname),
   },
 }
 
-export default withPayload(nextConfig)
+export default withPayload(nextConfig, { devBundleServerPackages: false })
