+++ templates/with-vercel-website/redirects.ts
@@ -1,9 +1,11 @@
-const redirects = async () => {
+import type { NextConfig } from 'next'
+
+export const redirects: NextConfig['redirects'] = async () => {
   const internetExplorerRedirect = {
     destination: '/ie-incompatible.html',
     has: [
       {
-        type: 'header',
+        type: 'header' as const,
         key: 'user-agent',
         value: '(.*Trident.*)', // all ie browsers
       },
@@ -12,9 +14,5 @@ const redirects = async () => {
     source: '/:path((?!ie-incompatible.html$).*)', // all pages except the incompatibility page
   }
 
-  const redirects = [internetExplorerRedirect]
-
-  return redirects
+  return [internetExplorerRedirect]
 }
-
-export default redirects
